GHC.TypeLits.Witnesses

Description

This module essentially provides a lightweight subset of the singletons library specifically for Nat and Symbol, from GHC.TypeLits.

Its main functionality is for first-class manipulation of KnownNat and KnownSymbol constraints. For example, in general, if you have KnownNat n, GHC can't infer KnownNat (n + 1). And, if you have both KnownNat n and KnownNat m, GHC can't infer 'KnownNat (n + m).

This can be annoying when dealing with libraries and applications where one regularly adds and subtracts type-level nats and expects KnownNat instances to follow. For example, vector concatenation of length-encoded vector types can be:

concat :: (KnownNat n, KnownNat m) => Vector n a -> Vector m a -> Vector (n + m) a

But, now n + m does not have a KnownNat instance...which makes operations like this much less useful.

Usually, the easiest way to get around this is to use a typechecker plugin, like https://hackage.haskell.org/package/ghc-typelits-knownnat. However, we can do this without the help of a typechecker plugin using first-class values, at the cost of some increased verbosity.

We introduce SNat n, which is a term-level witness of knownnat-ness that can be manipulated as a first-class value.

If we have KnownNat n, we can construct an SNat n:

SNat :: KnownNat n -> SNat n

Furthermore, if we have an SNat n, we can pattern match on the SNat constructor to get a KnownNat n constraint:

myFunc :: SNat n -> Bool
myFunc SNat = ...  -- in this body, we have `KnownNat n`

So if we have KnownNat n and KnownNat m, we can get KnownNat (n + m) by using %+, which adds together SNats:

case (SNat :: SNat n) %+ (SNat :: SNat m) of
  SNat -> -- in this branch, we have `KnownNat (n + m)`

Note that this module converts between SNat and Natural, and not SNat and Integer, in GHC.TypeNats-style.

Of course, all of this functionality is provided by the singletons library, in Data.Singletons.TypeLits. This module can be useful if you want a lightweight alternative without the full might of singletons. The main benefit of the singletons library is providing a unified interface for singletons of all different kinds/types, and not just Natural and String.

Nats

data SNat n

An SNat n is a witness for KnownNat n.

This means that if you pattern match on the SNat constructor, in that branch you will have a KnownNat n constraint.

myFunc :: SNat n -> Bool
myFunc SNat = ...  -- in this body, we have `KnownNat n`

This is essentially a singleton for Nat, and stands in for the singletons SNat and Sing types.

Constructors

KnownNat n => SNat

data SomeNat where

This type represents unknown type-level natural numbers.

Since: base-4.10.0.0

Bundled Patterns

pattern SomeNat_ :: SNat n -> SomeNat
A useful pattern synonym for matching on a SomeNat as if it contained an SNat n, and not a Proxy n as it exists in GHC.TypeLits.

A layer of compatibility letting us use the original SomeNat type in a way that works well with SNat.

This stands in for the singletons SomeSing constructor.

data Natural where

Type representing arbitrary-precision non-negative integers.

>>> 2^100 :: Natural
1267650600228229401496703205376

Operations whose result would be negative throw (Underflow :: ArithException),

>>> -1 :: Natural
*** Exception: arithmetic underflow

Since: base-4.8.0.0

Bundled Patterns

pattern FromSNat :: SNat n -> Natural
A useful pattern synonym for matching on a Natural as if it "were" a SNat:

myFunc :: Natural -> Bool
myFunc (FromSNat x) = ...  -- x is `SNat n`, with n coming from the input

It can be used as a function, as well, to convert an SNat n back into the Natural that it represents.

This stands in for the singletons FromSing pattern synonym.

fromSNat :: SNat n -> Natural

Convert ("reflect") an SNat back into the Natural it represents.

This stands in the singletons fromSing function.

withKnownNat :: SNat n -> (KnownNat n => r) -> r

Given an SNat n and a value that would require a KnownNat n instance, create that value.

This stands in for the function of the same name from Data.Singletons.TypeLits.

withSomeNat :: Natural -> (forall n. SNat n -> r) -> r

Promote ("reify") a Natural to an SNat n, by providing a continuation that would handle it in a way that is polymorphic over all possible n.

This stands in the singletons withSomeSing function.

toSomeNat :: Natural -> SomeNat

Promote ("reify") a Natural to an SNat n existentially hidden inside a SomeNat. To use it, pattern match using SomeNat_.

This stands in the singletons toSomeSing function.

Operations

(%+) :: SNat n -> SNat m -> SNat (n + m)

Addition of SNats.

This also will provide the correct KnownNat instance for SNat (n + m), so can be used as a way to "add" KnownNat instances.

This stands in for the function with the same name from Data.Singletons.Prelude.Num.

(%-) :: SNat n -> SNat m -> SNat (n - m)

Subtraction of SNats. Note that this is unsafe, as will trigger a run-time underflow if m is bigger than n even though it will always succeed at compiletime.

This also will provide the correct KnownNat instance for SNat (n - m), so can be used as a way to "subtract" KnownNat instances.

This stands in for the function with the same name from Data.Singletons.Prelude.Num.

minusSNat :: SNat n -> SNat m -> Either (CmpNat n m :~: LT) (SNat (n - m))

A safe version of %-: it will return Left if n is less than m (with a witness that it is), or else return the subtracted SNat in Right in a way that is guarunteed to not have runtime underflow.

minusSNat_ :: SNat n -> SNat m -> Maybe (SNat (n - m))

A version of minusSNat that just returns a Maybe.

(%*) :: SNat n -> SNat m -> SNat (n - m)

Addition of SNats.

This also will provide the correct KnownNat instance for SNat (n * m), so can be used as a way to "multiply" KnownNat instances.

This stands in for the function with the same name from Data.Singletons.Prelude.Num.

(%^) :: SNat n -> SNat m -> SNat (n ^ m)

Exponentiation of SNats.

This also will provide the correct KnownNat instance for SNat (n ^ m), so can be used as a way to "exponentiate" KnownNat instances.

This stands in for the function with the same name from Data.Singletons.TypeLits.

Compare

(%<=?) :: SNat n -> SNat m -> n :<=? m

Compare n and m, categorizing them into one of the constructors of :<=?.

sCmpNat :: SNat n -> SNat m -> SCmpNat n m

Compare n and m, categorizing them into one of the constructors of SCmpNat.

Unsafe

unsafeLiftNatOp1 :: (Natural -> Natural) -> SNat n -> SNat m

Lift a unary operation to act on an SNat n that returns an SNat m. The function given must properly describe the relationship between n and m.

For example:

double :: SNat n -> SNat (n * 2)
double = unsafeLiftNatOp1 (*2)

The correctness of the relationship is not checked, so be aware that this can cause programs to break.

unsafeLiftNatOp2 :: (Natural -> Natural -> Natural) -> SNat n -> SNat m -> SNat o

Lift a binary operation to act on an SNat n and SNat m that returns an SNat o. The function given must properly describe the relationship between n, m, and o.

For example:

multiply :: SNat n -> SNat m -> SNat (n * m)
multiply = unsafeLiftNatOp2 (*)

The correctness of the relationship is not checked, so be aware that this can cause programs to break.

Symbols

data SSymbol n

An SSymbol n is a witness for KnownSymbol n.

This means that if you pattern match on the SSymbol constructor, in that branch you will have a KnownSymbol n constraint.

myFunc :: SSymbol n -> Bool
myFunc SSymbol = ...  -- in this body, we have `KnownSymbol n`

This is essentially a singleton for Symbol, and stands in for the singletons SSymbol and Sing types.

Constructors

KnownSymbol n => SSymbol

data SomeSymbol where

This type represents unknown type-level symbols.

Bundled Patterns

pattern SomeSymbol_ :: SSymbol n -> SomeSymbol
A useful pattern synonym for matching on a SomeSymbol as if it contained an SSymbol n, and not a Proxy n as it exists in GHC.TypeLits.

A layer of compatibility letting us use the original SomeSymbol type in a way that works well with SSymbol.

This stands in for the singletons SomeSing constructor.

pattern FromSSymbol :: SSymbol n -> String

A useful pattern synonym for matching on a String as if it "were" a SSymbol:

myFunc :: String -> Bool
myFunc (FromSSymbol x) = ...  -- x is `SSymbol n`, with n coming from the input

It can be used as a function, as well, to convert an SSymbol n back into the String that it represents.

This stands in for the singletons FromSing pattern synonym, except it matches on a String instead of a Text.

fromSSymbol :: SSymbol n -> String

Convert ("reflect") an SSymbol back into the String it represents.

This stands in the singletons fromSing function, except it returns a String instead of Text.

withKnownSymbol :: SSymbol n -> (KnownSymbol n => r) -> r

Given an SSymbol n and a value that would require a KnownSymbol n instance, create that value.

This stands in for the function of the same name from Data.Singletons.TypeLits.

withSomeSymbol :: String -> (forall n. SSymbol n -> r) -> r

Promote ("reify") a String to an SSymbol n, by providing a continuation that would handle it in a way that is polymorphic over all possible n.

This stands in the singletons withSomeSing function, except it takes a String instead of Text.

toSomeSymbol :: String -> SomeSymbol

Promote ("reify") a String to an SSymbol n existentially hidden inside a SomeNat. To use it, pattern match using SomeSymbol_.

This stands in the singletons toSomeSing function, except it takes a String instead of Text.